refactor: Merge types/docs from crud/response_conversion.pyi
- Merged type hints and docstrings from `crudclient/crud/response_conversion.pyi` into `crudclient/crud/response_conversion.py`.

- Deleted `crudclient/crud/response_conversion.pyi`.
